OM-43603,RB-29636 Market: Handle CostDTOs with Region information

+++ com.vmturbo.analysis/platform.analysis/src/main/java/com/vmturbo/platform/analysis/economy/BalanceAccount.java
@@ -9,9 +9,9 @@ public class BalanceAccount {
 
     private double spent_;
     private double budget_;
-    private long id_;
+    private int id_;
 
-    public BalanceAccount(double spent, double budget, long id) {
+    public BalanceAccount(double spent, double budget, int id) {
         spent_ = spent;
         budget_ = budget;
         id_ = id;
@@ -33,7 +33,7 @@ public class BalanceAccount {
         return budget_;
     }
 
-    public long getId() {
+    public int getId() {
         return id_;
     }
 }
